feat. 1차 검사 베포용 (키, 몸무게 ,float 변경)

+++ src/main/java/com/example/fitness_trAIner/controller/admin/dto/request/AdminUserUpdateRequestBody.java
@@ -11,8 +11,8 @@ import lombok.Getter;
 public class AdminUserUpdateRequestBody {
     private Long id;
     private String nickname;
-    private int height;
-    private int weight;
+    private float height;
+    private float weight;
     private int age;
     private int spicyPreference;
     private Boolean meatConsumption;
