+++ src/main/java/com/example/fitness_trAIner/service/user/dto/request/UserServiceSignupRequest.java
@@ -12,8 +12,8 @@ public class UserServiceSignupRequest {
     private String username;
     private String password;
     private String nickname;
-    private int height;
-    private int weight;
+    private float height;
+    private float weight;
     private int age;
     private int spicyPreference;
     private Boolean meatConsumption;
